Move lookup functions from RoomMemberHandler to IdentityHandler
